+++ packages/f-content-cards/stories/ContentCards.stories.js
@@ -1,31 +1,102 @@
-import { withKnobs, array, text } from '@storybook/addon-knobs';
+import {
+    withKnobs,
+    optionsKnob as options,
+    text,
+    button
+} from '@storybook/addon-knobs';
+import { withA11y } from '@storybook/addon-a11y';
+import mock, { proxy } from 'xhr-mock';
+
 import ContentCards from '../src/components/ContentCards.vue';
 import { defaultEnabledCardTypes } from '../src/services/contentCard.service';
-import { withA11y } from '@storybook/addon-a11y';
+
+import cards from './mockData/cards';
+import data from './mockData/data';
+
+/**
+ * Resets all locally stored braze data so that the stubbed data is always fresh on page load
+ */
+function resetBrazeData () {
+    document.cookie
+        .split('; ')
+        .filter(row => row.startsWith('ab.'))
+        .map(row => row.split('=')[0])
+        .forEach(cookieName => {
+            document.cookie = `${cookieName}=;max-age=0`;
+        });
+
+    Object.keys(localStorage)
+        .filter(row => row.startsWith('ab.'))
+        .forEach(storageItem => {
+            localStorage.removeItem(storageItem);
+        });
+}
+
+const allowedCardTypes = {
+    'Terms and Conditions': 'Terms_And_Conditions_Card',
+    'Terms and Conditions 2': 'Terms_And_Conditions_Card_2',
+    Header: 'Header_Card',
+    Voucher: 'Voucher_Card_1',
+    Recommendation: 'Recommendation_Card_1',
+    'Promotion 1': 'Promotion_Card_1',
+    'Promotion 2': 'Promotion_Card_2',
+    'Home Promotion 1': 'Home_Promotion_Card_1',
+    'Home Promotion 2': 'Home_Promotion_Card_2',
+    'Post Order 1': 'Post_Order_Card_1',
+    'Anniversary 1': 'Anniversary_Card_1',
+    'Restaurant FTC Offer': 'Restaurant_FTC_Offer_Card',
+    'Recommendation 1': 'Recommendation_Card_1'
+};
 
 export default {
     title: 'Components/Organisms',
     decorators: [withKnobs, withA11y]
 };
 
-export const ContentCardscomponent = () => ({
-    components: { ContentCards },
-    props: {
-        apiKey: {
-            default: text('API Key', '')
+export function ContentCardscomponent () {
+    return {
+        components: {
+            ContentCards
         },
-        userId: {
-            default: text('User ID', '')
+
+        props: {
+            apiKey: {
+                default: text('API Key', '00000000-0000-0000-0000-000000000000')
+            },
+            userId: {
+                default: text('User ID', 'test-user-id')
+            },
+            title: {
+                default: text('Title', 'Promotional Offers')
+            },
+            enabledCardTypes: {
+                default: options('Enabled Card Types', allowedCardTypes, defaultEnabledCardTypes, {
+                    display: 'multi-select'
+                })
+            },
+            refreshDisplayedCards: {
+                default: button('Refresh Card Types Displayed', () => { window.appboy.requestContentCardsRefresh(); })
+            }
         },
-        title: {
-            default: text('Title', 'Promotional Offers')
+
+        beforeCreate: () => {
+            resetBrazeData();
+            mock.teardown();
+            mock.setup();
+            mock.post(/\/api\/v3\/content_cards\/sync\/?/, {
+                status: 201,
+                body: JSON.stringify(cards())
+            });
+            mock.post(/\/api\/v3\/data\/?/, {
+                status: 201,
+                body: JSON.stringify(data())
+            });
+            mock.use(proxy);
         },
-        enabledCardTypes: {
-            default: array('Enabled Card Types', defaultEnabledCardTypes)
-        }
-    },
-    template: '<content-cards :userId="userId" :apiKey="apiKey" :title="title" :enabledCardTypes="enabledCardTypes" />'
-});
+
+        template: '<content-cards :userId="userId" :apiKey="apiKey" :title="title" :enabledCardTypes="enabledCardTypes" />'
+    };
+}
 
 ContentCardscomponent.story = {
     name: 'f-content-cards'